pyproject.toml
[project]
name = "emlo4-session-04-ajithvcoder"
version = "0.1.0"
description = "Use lightning to train, eval and infer a dataset"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=24.8.0",
    "lightning[extra]==2.4.0",
    "loguru>=0.7.2",
    "rich>=13.8.1",
    "tensorboard==2.17.1",
    "timm==1.0.9",
    "torch==2.4.1+cpu",
    "torchvision==0.19.1+cpu",
    "gdown==5.2.0",
    "hydra-core==1.3.0",
    "python-dotenv",
    "rootutils"
]

[tool.coverage.report]
include = ["src/**/*.py"]
show_missing = true